Main page, correct base, and minor improvements (#8)
Create main page design and content, correct base and add dropdown menu there. Also rename container in authorization to not conflict with the one in footer and header. + minor fixes

* adding to settings several statifiles directories, create 1 static directory

* creating page for clubs, adding clubs to the main page

* move static dir to the other place

* created model for clubs, add css for page of clubs, some backend in urls,views

* added pages for resetting password, also added some settings in settings.py for resetting password

* added links in urls.py in authorization, added style for form reset

* added model clubs in the settings

* rewriting the database

* Removed printing the key on server launch

* Fix typo in base.html

* Main page first revision
Add club types to the main page, format files with autopep8, add
model of club type with sample entries

* Add upcoming events and fix base.html

* Add dropdown menu. Fix register and login pages.
Dropdown menu is added to base (as it is shown in every page).
Register and login pages are now displayed correctly together with
header and footer.

* Remove commented lines

* Remove more commented lines

* Shorten lines to comply with linter

* Add settings for linter

Co-authored-by: Dablup <d.livitin@innopolis.university>